backend/recommender/requirements.txt
lightfm
pandas
scipy
pymongo
python-dotenv
numpy